working with ble devkits, unlocks door when close
*NOTE: This update was manually applied on github because my local github wasn't recognizing changes in this file. 

A lot of things fell into place here. The serial read is working, decoding the cobs message and stripping out the relavent data: ID and RSSI. Note that the ID is currently assuming its a 4 byte unsigned long in network byte order. It also assumes that those 4 bytes starts at the 5 byte and ends with the 8th (inclusive).

Those values are currently being passed with the update function, which looks up the ID in the white-list and, if needed, it records that new value for that ID. This is used to calculate the running average RSSI.

Once the average reaches a threshold constant of "-20", the door unlock subroutine is called. Currently it sends a COBS frame back down the serial port, the contents of that frame are based on previous work we did with opening doors and serial comms.

At this point it is as done as it can be until we decide on specific serial wire line protocols for incoming and outgoing messages.